+++ crates/account_sdk/src/tests/mod.rs
@@ -5,3 +5,4 @@ pub mod runners;
 mod runners_test;
 mod stark_pair_auth_test;
 mod webauthn;
+mod session;
